agentic-ai-workshop/ai-agent-demo/requirements.txt
openai
mcp
python-dotenv
langchain
langchain-openai
langchain-mongodb
pymongo
certifi
fastapi
uvicorn
langchain-text-splitters
python-multipart
pypdf
python-docx
